backend/pyproject.toml
[project]
name = "workout-tracker-api"
version = "0.1.0"
description = "FastAPI backend for Workout Tracker"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "SQLAlchemy>=2.0.30",
  "psycopg[binary]>=3.1.19",
  "alembic>=1.13.2",
  "passlib[bcrypt]>=1.7.4",
  "PyJWT>=2.8.0",
  "python-multipart>=0.0.9",
  "email-validator>=2.1.1"
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
